+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import {Box} from "@mui/material";
-import {Routes, Route} from "react-router-dom";
+import {Routes, Route, Navigate, useNavigate} from "react-router-dom";
 
 import Layout from "@layout/Layout";
 import Home from "@pages/Home";
@@ -14,8 +14,12 @@ import SoonGraph from "@pages/Soon/SoonGraph";
 import Admin from "@pages/Admin/Admin";
 import MyProfile from "@pages/MyProfile/MyProfile";
 import Withdrawal from "@pages/Withdrawal/Withdrawal";
+import {userGoogleAuthState} from "@recoils/Login/state";
+import {useRecoilValue} from "recoil";
 
 export default function App() {
+  const googleAuth = useRecoilValue(userGoogleAuthState);
+  // const navigate = useNavigate();
   return (
     <Box>
       <Routes>
@@ -27,7 +31,9 @@ export default function App() {
 
           {/*Not Login */}
           <Route path="/login" element={Auth(Login, null)}></Route>
-          <Route path="/register" element={Auth(Register, null)}></Route>
+          <Route
+            path="/register"
+            element={googleAuth == null || googleAuth?.status != "REGISTER" ? <Navigate to={"/"} /> : Auth(Login, null)}></Route>
           <Route path="/register/:userid" element={Auth(Register, null)}></Route>
           <Route path="/history" element={Auth(HistoryWrite, null)}></Route>
           <Route path="/history/:historyid" element={Auth(HistoryWrite, null)}></Route>
