+++ frontend/src/components/pages/Login/Login.tsx
@@ -13,21 +13,22 @@ const Login = () => {
 
   const navigate = useNavigate();
   const handleLoginSuccess = async (code: string) => {
-    const {data} = await getGoogleInfoAxios(code);
-    const status = data?.status;
+    const {
+      data: {status, auth, user},
+    } = await getGoogleInfoAxios(code);
+    // const status = data?.status;
     if (status == "REGISTER") {
-      setGoogleAuth(data?.auth);
+      setGoogleAuth(Object.assign(auth, status));
 
       //TODO: 회원 가입 폼 이동
       console.log("회원 가입하시죠");
       navigate("/register");
     } else {
-      console.log("로그인 process 진행하시죠", data);
-      const {ssoid} = data?.auth;
-      const {userid} = data?.user;
-      const user = data?.user;
+      console.log("로그인 process 진행하시죠", auth);
+      const {ssoid} = auth;
+      const {userid} = user;
       const result = await axios.post("/auth/token", {userid, ssoid});
-      console.log("result : ", result?.data);
+      // console.log("result : ", result?.data);
       setUser(user); // loginUser, #user 통채로 저장하지 않고, access_token으로 가져오도록 수정
       setGoogleAuth(null); //혹시 들어잇을지 모르니 지운다
       navigate("/");
